# Conflicts:
#	modules/networker/.gitignore
#	modules/networker/build.gradle
#	modules/networker/inet.uniq
#	modules/networker/ostpst/build.gradle
#	modules/networker/src/main/java/ru/vachok/networker/AppComponents.java
#	modules/networker/src/main/java/ru/vachok/networker/AppInfoOnLoad.java
#	modules/networker/src/main/java/ru/vachok/networker/ConstantsFor.java
#	modules/networker/src/main/java/ru/vachok/networker/DBPropsCallable.java
#	modules/networker/src/main/java/ru/vachok/networker/IntoApplication.java
#	modules/networker/src/main/java/ru/vachok/networker/TForms.java
#	modules/networker/src/main/java/ru/vachok/networker/abstr/Pinger.java
#	modules/networker/src/main/java/ru/vachok/networker/accesscontrol/MatrixSRV.java
#	modules/networker/src/main/java/ru/vachok/networker/accesscontrol/NameOrIPChecker.java
#	modules/networker/src/main/java/ru/vachok/networker/accesscontrol/common/CommonSRV.java
#	modules/networker/src/main/java/ru/vachok/networker/accesscontrol/inetstats/InetStatSorter.java
#	modules/networker/src/main/java/ru/vachok/networker/accesscontrol/inetstats/InetUserPCName.java
#	modules/networker/src/main/java/ru/vachok/networker/accesscontrol/sshactions/Tracerouting.java
#	modules/networker/src/main/java/ru/vachok/networker/ad/user/ADUser.java
#	modules/networker/src/main/java/ru/vachok/networker/ad/user/MoreInfoWorker.java
#	modules/networker/src/main/java/ru/vachok/networker/componentsrepo/Visitor.java
#	modules/networker/src/main/java/ru/vachok/networker/controller/ActDirectoryCTRL.java
#	modules/networker/src/main/java/ru/vachok/networker/controller/AnketaKonfetaCRTL.java
#	modules/networker/src/main/java/ru/vachok/networker/controller/CalculateCTRL.java
#	modules/networker/src/main/java/ru/vachok/networker/controller/MatrixCtr.java
#	modules/networker/src/main/java/ru/vachok/networker/controller/NetScanCtr.java
#	modules/networker/src/main/java/ru/vachok/networker/controller/ServiceInfoCtrl.java
#	modules/networker/src/main/java/ru/vachok/networker/controller/SshActsCTRL.java
#	modules/networker/src/main/java/ru/vachok/networker/exe/ThreadConfig.java
#	modules/networker/src/main/java/ru/vachok/networker/exe/runnabletasks/ExecScan.java
#	modules/networker/src/main/java/ru/vachok/networker/exe/runnabletasks/NetMonitorPTV.java
#	modules/networker/src/main/java/ru/vachok/networker/exe/runnabletasks/NetScannerSvc.java
#	modules/networker/src/main/java/ru/vachok/networker/exe/runnabletasks/PfListsSrv.java
#	modules/networker/src/main/java/ru/vachok/networker/exe/runnabletasks/ScanOnline.java
#	modules/networker/src/main/java/ru/vachok/networker/exe/runnabletasks/SpeedChecker.java
#	modules/networker/src/main/java/ru/vachok/networker/exe/runnabletasks/TemporaryFullInternet.java
#	modules/networker/src/main/java/ru/vachok/networker/exe/runnabletasks/external/SaveLogsToDB.java
#	modules/networker/src/main/java/ru/vachok/networker/exe/schedule/DiapazonScan.java
#	modules/networker/src/main/java/ru/vachok/networker/fileworks/DeleterTemp.java
#	modules/networker/src/main/java/ru/vachok/networker/fileworks/FileSearcher.java
#	modules/networker/src/main/java/ru/vachok/networker/fileworks/FileSystemWorker.java
#	modules/networker/src/main/java/ru/vachok/networker/net/AccessListsCheckUniq.java
#	modules/networker/src/main/java/ru/vachok/networker/net/NetListKeeper.java
#	modules/networker/src/main/java/ru/vachok/networker/net/NetPinger.java
#	modules/networker/src/main/java/ru/vachok/networker/net/NetScanFileWorker.java
#	modules/networker/src/main/java/ru/vachok/networker/net/PCUserResolver.java
#	modules/networker/src/main/java/ru/vachok/networker/net/SwitchesAvailability.java
#	modules/networker/src/main/java/ru/vachok/networker/net/TestServer.java
#	modules/networker/src/main/java/ru/vachok/networker/net/enums/ConstantsNet.java
#	modules/networker/src/main/java/ru/vachok/networker/net/enums/OtherKnownDevices.java
#	modules/networker/src/main/java/ru/vachok/networker/net/libswork/CoverReportUpdate.java
#	modules/networker/src/main/java/ru/vachok/networker/net/libswork/RegRuFTPLibsUploader.java
#	modules/networker/src/main/java/ru/vachok/networker/services/ADSrv.java
#	modules/networker/src/main/java/ru/vachok/networker/services/AnketaKonfeta.java
#	modules/networker/src/main/java/ru/vachok/networker/services/TimeChecker.java
#	modules/networker/src/main/java/ru/vachok/networker/services/WhoIsWithSRV.java
#	modules/networker/src/main/java/ru/vachok/networker/statistics/InteretStats.java
#	modules/networker/src/main/java/ru/vachok/networker/systray/SystemTrayHelper.java
#	modules/networker/src/main/resources/templates/common.html
#	modules/networker/src/main/resources/templates/ping.html
#	modules/networker/src/main/resources/templates/vir.html
#	modules/networker/src/test/java/ru/vachok/networker/AppComponentsTest.java
#	modules/networker/src/test/java/ru/vachok/networker/AppInfoOnLoadTest.java
#	modules/networker/src/test/java/ru/vachok/networker/DBPropsCallableTest.java
#	modules/networker/src/test/java/ru/vachok/networker/ExitAppTest.java
#	modules/networker/src/test/java/ru/vachok/networker/IntoApplicationTest.java
#	modules/networker/src/test/java/ru/vachok/networker/accesscontrol/inetstats/InetIPUserTest.java
#	modules/networker/src/test/java/ru/vachok/networker/accesscontrol/inetstats/InetUserPCNameTest.java
#	modules/networker/src/test/java/ru/vachok/networker/accesscontrol/sshactions/TraceroutingTest.java
#	modules/networker/src/test/java/ru/vachok/networker/exe/runnabletasks/ScanOnlineTest.java
#	modules/networker/src/test/java/ru/vachok/networker/exe/schedule/DiapazonScanTest.java
#	modules/networker/src/test/java/ru/vachok/networker/net/TelnetStarterTest.java
#	modules/networker/src/test/java/ru/vachok/networker/net/libswork/CoverReportUpdateTest.java
#	modules/networker/src/test/java/ru/vachok/networker/net/libswork/RegRuFTPLibsUploaderTest.java
#	modules/networker/src/test/java/ru/vachok/networker/services/SpeedCheckerTest.java
#	modules/networker/src/test/java/ru/vachok/networker/statistics/InteretStatsTest.java

+++ modules/networker/src/main/java/ru/vachok/networker/statistics/InteretStats.java
@@ -7,7 +7,6 @@ import com.mysql.jdbc.jdbc2.optional.MysqlDataSource;
 import ru.vachok.messenger.MessageSwing;
 import ru.vachok.messenger.MessageToUser;
 import ru.vachok.mysqlandprops.RegRuMysql;
-import ru.vachok.networker.AppComponents;
 import ru.vachok.networker.ConstantsFor;
 import ru.vachok.networker.TForms;
 import ru.vachok.networker.abstr.DataBaseRegSQL;
@@ -18,8 +17,12 @@ import ru.vachok.networker.services.MessageLocal;
 import java.awt.*;
 import java.io.*;
 import java.sql.*;
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
 import java.time.LocalTime;
+import java.util.Date;
 import java.util.List;
+import java.util.concurrent.Executors;
 
 
 /**
@@ -31,15 +34,54 @@ public class InteretStats implements Runnable, DataBaseRegSQL {
     
     private static final String FILENAME_INETSTATSCSV = "inetstats.csv";
     
-    private static final String SQL_DISTINCTIPSWITHINET = "SELECT DISTINCT `ip` FROM `inetstats`";
+    private String fileName = "null";
     
     private Connection connectionF;
     
     private Savepoint savepoint;
     
-    private String fileName;
+    private String sql = "null";
     
-    private String sql;
+    private static final String SQL_DISTINCTIPSWITHINET = ConstantsFor.SQL_SELECTINETSTATS;
+    
+    /**
+     Для тестов
+     <p>
+     
+     @return {@link #fileName}
+     */
+    protected String getFileName() {
+        return fileName;
+    }
+    
+    /**
+     Для тестов
+     <p>
+     
+     @param fileName имя csv-файла
+     */
+    protected void setFileName(String fileName) {
+        this.fileName = fileName;
+    }
+    
+    /**
+     For tests
+     
+     @return {@link #sql}
+     */
+    protected String getSql() {
+        return sql;
+    }
+    
+    /**
+     Для тестов
+     <p>
+     
+     @param sql sql-запрос
+     */
+    protected void setSql(String sql) {
+        this.sql = sql;
+    }
     
     private MessageToUser messageToUser = new MessageLocal(getClass().getSimpleName());
     
@@ -53,7 +95,7 @@ public class InteretStats implements Runnable, DataBaseRegSQL {
         long iPsWithInet = readIPsWithInet();
         messageToUser.info(getClass().getSimpleName() + "in kbytes. ", new File(FILENAME_INETSTATSIPCSV).getAbsolutePath(), " = " + iPsWithInet);
         readInetStatsRSetToCSV();
-        AppComponents.threadConfig().execByThreadConfig(new InetStatSorter());
+        Executors.unconfigurableExecutorService(Executors.newSingleThreadExecutor()).execute(new InetStatSorter());
     }
     
     @Override public int selectFrom() {
@@ -148,6 +190,8 @@ public class InteretStats implements Runnable, DataBaseRegSQL {
     
     private void readInetStatsRSetToCSV() {
         List<String> chkIps = FileSystemWorker.readFileToList(new File(FILENAME_INETSTATSIPCSV).getPath());
+        DateFormat format = new SimpleDateFormat("E");
+        String weekDay = format.format(new Date());
         long totalBytes = 0;
         for (String ip : chkIps) {
             this.fileName = FILENAME_INETSTATSCSV.replace("inetstats", ip).replace(".csv", "_" + LocalTime.now().toSecondOfDay() + ".csv");
@@ -157,7 +201,7 @@ public class InteretStats implements Runnable, DataBaseRegSQL {
             totalBytes += file.length();
             new MessageLocal(getClass().getSimpleName())
                 .info(fileName, file.length() / ConstantsFor.KBYTE + " kb", "total kb: " + totalBytes / ConstantsFor.KBYTE);
-            if (file.length() > 10) {
+            if (weekDay.equals("вс") && file.length() > 10) {
                 this.sql = new StringBuilder().append("DELETE FROM `inetstats` WHERE `ip` LIKE '").append(ip).append("'").toString();
                 System.out.println(deleteFrom() + " rows deleted.");
             }
